study_elsticsearch/requirements.txt
Django==1.11.16
django-haystack==2.8.1
djangorestframework==3.9.4
drf-haystack==1.8.5
elasticsearch==2.4.1
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
urllib3==1.25.6
